added testcases for select_pokemon
