--- test_luckypokemon.py
@@ -1,19 +1,0 @@
-import pytest
-from pypokemon import luckypokemon
-def test_luckypokemon():
-    pokemon_list = [
-        "Pikachu", "Charmander", "Bulbasaur", "Squirtle", "Jigglypuff", 
-        "Meowth", "Psyduck", "Snorlax", "Eevee", "Gyarados", 
-        "Lapras", "Ditto", "Vaporeon", "Jolteon", "Flareon",
-        "Machamp", "Gengar", "Dragonite", "Mewtwo", "Mew",
-        "Togepi", "Espeon", "Umbreon", "Scizor", "Heracross",
-        "Blaziken", "Swampert", "Sceptile", "Gardevoir", "Salamence",
-        "Metagross", "Lucario", "Garchomp", "Infernape", "Torterra",
-        "Empoleon", "Roserade", "Glaceon", "Leafeon", "Togekiss",
-        "Zoroark", "Hydreigon", "Greninja", "Aegislash", "Talonflame",
-        "Sylveon", "Decidueye", "Incineroar", "Primarina", "Zacian"
-    ]
-    for i in range(100):
-        result = luckypokemon()
-        lucky_pokemon = result.split("... ")[1].split("!")[0]  # Extract Pokémon name
-        assert lucky_pokemon in pokemon_list
